Address gemini-code-assist nit: drop redundant np.float32 casts
Gemini Code Assist flagged the .astype(np.float32) calls on image and
blended_bg inside apply_soft_background_texture as redundant — tight_mask
is already float32 (it's focus_mask * focus_mask, where focus_mask is
float32 from full_face_mask construction in get_dynamic_region_masks).
NumPy auto-promotes uint8 * float32 → float32, so the explicit casts
allocate two unnecessary intermediate buffers per frame for no gain.

Same one-line fix applied to all 4 algorithm files (V9 + V10, both
Windows + macOS) for full platform parity:

  out = (image.astype(np.float32) * tight_mask
         + blended_bg.astype(np.float32) * inverse_mask)
  →
  out = image * tight_mask + blended_bg * inverse_mask

Behavior is bit-for-bit identical (the auto-promotion produces the same
float32 result as the explicit cast); the only difference is two fewer
intermediate allocations per frame, which is meaningful at 30 FPS.

427/427 tests pass.

macOS compat: the macOS copies (oldcam-v9/macOS/oldcam.py,
oldcam-v10/macOS/oldcam.py) get the identical change so V9/V10 stay
visually consistent across platforms.

+++ oldcam-v9/oldcam.py
@@ -547,7 +547,12 @@ def apply_soft_background_texture(image, focus_mask, strength=0.08):
     small = cv2.resize(image, (max(1, w // 3), max(1, h // 3)), interpolation=cv2.INTER_LINEAR)
     restored = cv2.resize(small, (w, h), interpolation=cv2.INTER_LINEAR)
     blended_bg = cv2.addWeighted(image, 1.0 - strength, restored, strength, 0)
-    out = (image.astype(np.float32) * tight_mask + blended_bg.astype(np.float32) * inverse_mask)
+    # Explicit .astype(np.float32) on image/blended_bg is redundant — tight_mask
+    # is already float32 (focus_mask is float32 from full_face_mask construction
+    # in get_dynamic_region_masks), so NumPy auto-promotes uint8 * float32 →
+    # float32. Saves two unnecessary intermediate allocations per frame
+    # (gemini-code-assist nit on PR #20).
+    out = image * tight_mask + blended_bg * inverse_mask
     return np.clip(out, 0, 255).astype(np.uint8)
